+++ packages/private/testing/index.ts
@@ -6,4 +6,4 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-export * from './src/render3';
+export * from './src/utils';
